PHP-8.2 support (#1800)
