+++ docker-compose.yml
@@ -57,6 +57,7 @@ services:
   '8.0-buster': { <<: *base_php_service, image: 'datadog/dd-trace-ci:php-8.0_buster' }
   '8.0-buster-shared-ext': { <<: *base_php_service, image: 'datadog/dd-trace-ci:php-8.0-shared-ext' }
   '8.1-buster': { <<: *base_php_service, image: 'datadog/dd-trace-ci:php-8.1_buster' }
+  '8.2-buster': { <<: *base_php_service, image: 'datadog/dd-trace-ci:php-8.2_buster' }
   'php-master-buster': { <<: *base_php_service, image: 'datadog/dd-trace-ci:php-master_buster' }
   # --- CentOS 6 ---
   '7.0-centos7': { <<: *base_php_service, image: 'datadog/dd-trace-ci:php-7.0_centos-7' }
@@ -66,6 +67,7 @@ services:
   '7.4-centos7': { <<: *base_php_service, image: 'datadog/dd-trace-ci:php-7.4_centos-7' }
   '8.0-centos7': { <<: *base_php_service, image: 'datadog/dd-trace-ci:php-8.0_centos-7' }
   '8.1-centos7': { <<: *base_php_service, image: 'datadog/dd-trace-ci:php-8.1_centos-7' }
+  '8.2-centos7': { <<: *base_php_service, image: 'datadog/dd-trace-ci:php-8.2_centos-7' }
 
   mysql_integration:
     image: mysql/mysql-server:8.0
