+++ dockerfiles/ci/buster/docker-compose.yml
@@ -11,6 +11,17 @@ services:
           - linux/arm64
           - linux/amd64
 
+  php-8.2:
+    image: datadog/dd-trace-ci:php-8.2_buster
+    build:
+      context: .
+      dockerfile: php-8.2/Dockerfile
+      x-bake: *bake
+      args:
+        phpVersion: "8.2"
+        phpTarGzUrl: https://github.com/php/php-src/archive/a814afb08823bb0cdf4ae45f67a7db83c8be9cd7.tar.gz
+        phpSha256Hash: 584d2925889a7d388084150865fa238634043f38f2e060c27ffa8ac4d682d5ec
+
   php-8.1:
     image: datadog/dd-trace-ci:php-8.1_buster
     build:
