+++ dockerfiles/ci/centos/7/docker-compose.yml
@@ -87,3 +87,14 @@ services:
         phpTarGzUrl: https://www.php.net/distributions/php-8.1.8.tar.gz
         phpSha256Hash: 889d910558d2492f7f2236921b9bcde620674c8b684ec02d126060f8ca45dc8d
     image: 'datadog/dd-trace-ci:php-8.1_centos-7'
+
+  php-8.2:
+    build:
+      context: .
+      dockerfile: php.Dockerfile
+      x-bake: *bake
+      args:
+        phpVersion: "8.2"
+        phpTarGzUrl: https://github.com/php/php-src/archive/a814afb08823bb0cdf4ae45f67a7db83c8be9cd7.tar.gz
+        phpSha256Hash: 584d2925889a7d388084150865fa238634043f38f2e060c27ffa8ac4d682d5ec
+    image: 'datadog/dd-trace-ci:php-8.2_centos-7'
