+++ dockerfiles/ci/xfail_tests/8.2.list
@@ -1,0 +1,166 @@
+Zend/tests/fibers/out-of-memory-in-fiber.phpt
+Zend/tests/fibers/out-of-memory-in-nested-fiber.phpt
+Zend/tests/fibers/out-of-memory-in-recursive-fiber.phpt
+Zend/tests/014.phpt
+Zend/tests/arginfo_zpp_mismatch.phpt
+Zend/tests/arginfo_zpp_mismatch_strict.phpt
+Zend/tests/bug63882_2.phpt
+Zend/tests/bug69315.phpt
+Zend/tests/bug70258.phpt
+Zend/tests/bug70805.phpt
+Zend/tests/bug70805_1.phpt
+Zend/tests/bug70805_2.phpt
+Zend/tests/bug71539_5.phpt
+Zend/tests/bug72038.phpt
+Zend/tests/bug75921.phpt
+Zend/tests/bug78182.phpt
+Zend/tests/bug81104.phpt
+Zend/tests/gc_031.phpt
+Zend/tests/gc_037.phpt
+Zend/tests/gc_045.phpt
+Zend/tests/get_defined_functions_basic.phpt
+Zend/tests/get_required_files.phpt
+Zend/tests/gh7958.phpt
+Zend/tests/list_keyed_conversions.phpt
+Zend/tests/multibyte/multibyte_encoding_004.phpt
+Zend/tests/object_gc_in_shutdown.phpt
+Zend/tests/offset_array.phpt
+Zend/tests/strict_001.phpt
+Zend/tests/type_declarations/scalar_return_basic_64bit.phpt
+Zend/tests/type_declarations/scalar_strict_64bit.phpt
+Zend/tests/weakrefs/weakmap_basic_map_behavior.phpt
+Zend/tests/weakrefs/weakmap_error_conditions.phpt
+ext/curl/tests/bug45161.phpt
+ext/curl/tests/bug64267.phpt
+ext/curl/tests/bug76675.phpt
+ext/curl/tests/bug77535.phpt
+ext/curl/tests/bug79033.phpt
+ext/curl/tests/curl_int_cast.phpt
+ext/date/tests/bug52113.phpt
+ext/curl/tests/curl_postfields_array.phpt
+ext/fileinfo/tests/finfo_open_error.phpt
+ext/mbstring/tests/zend_multibyte-01.phpt
+ext/mbstring/tests/zend_multibyte-02.phpt
+ext/mbstring/tests/zend_multibyte-06.phpt
+ext/openssl/tests/bug46127.phpt
+ext/openssl/tests/bug48182.phpt
+ext/openssl/tests/bug54992.phpt
+ext/openssl/tests/bug65538_001.phpt
+ext/openssl/tests/bug65538_003.phpt
+ext/openssl/tests/bug65729.phpt
+ext/openssl/tests/bug68265.phpt
+ext/openssl/tests/bug68879.phpt
+ext/openssl/tests/bug68920.phpt
+ext/openssl/tests/bug69215.phpt
+ext/openssl/tests/bug72333.phpt
+ext/openssl/tests/bug74159.phpt
+ext/openssl/tests/bug77390.phpt
+ext/openssl/tests/capture_peer_cert_001.phpt
+ext/openssl/tests/openssl_encrypt_ccm.phpt
+ext/openssl/tests/openssl_peer_fingerprint_basic.phpt
+ext/openssl/tests/openssl_x509_checkpurpose_basic.phpt
+ext/openssl/tests/peer_verification.phpt
+ext/openssl/tests/san_peer_matching.phpt
+ext/openssl/tests/session_meta_capture.phpt
+ext/openssl/tests/session_meta_capture_tlsv13.phpt
+ext/openssl/tests/stream_crypto_flags_001.phpt
+ext/openssl/tests/stream_crypto_flags_002.phpt
+ext/openssl/tests/stream_crypto_flags_003.phpt
+ext/openssl/tests/stream_crypto_flags_004.phpt
+ext/openssl/tests/stream_security_level.phpt
+ext/openssl/tests/stream_server_reneg_limit.phpt
+ext/openssl/tests/stream_verify_peer_name_001.phpt
+ext/openssl/tests/stream_verify_peer_name_002.phpt
+ext/openssl/tests/stream_verify_peer_name_003.phpt
+ext/openssl/tests/streams_crypto_method.phpt
+ext/openssl/tests/tls_min_v1.0_max_v1.1_wrapper.phpt
+ext/openssl/tests/tls_wrapper.phpt
+ext/openssl/tests/tls_wrapper_with_tls_v1.3.phpt
+ext/openssl/tests/tlsv1.0_wrapper.phpt
+ext/openssl/tests/tlsv1.1_wrapper.phpt
+ext/openssl/tests/tlsv1.2_wrapper.phpt
+ext/openssl/tests/tlsv1.3_wrapper.phpt
+ext/pcntl/tests/pcntl_unshare_01.phpt
+ext/pcntl/tests/pcntl_unshare_02.phpt
+ext/pcntl/tests/pcntl_unshare_03.phpt
+ext/pdo_sqlite/tests/bug43831.phpt
+ext/pdo_sqlite/tests/bug44327_2.phpt
+ext/pdo_sqlite/tests/bug44327_3.phpt
+ext/pdo_sqlite/tests/bug48773.phpt
+ext/pdo_sqlite/tests/pdo_fetch_func_001.phpt
+ext/pdo_sqlite/tests/pdo_sqlite_lastinsertid.phpt
+ext/pdo/tests/pdo_023.phpt
+ext/pdo/tests/pdo_030.phpt
+ext/phar/tests/031.phpt
+ext/phar/tests/032.phpt
+ext/phar/tests/fatal_error_webphar.phpt
+ext/phar/tests/phar_buildfromiterator10.phpt
+ext/phar/tests/phar_buildfromiterator8.phpt
+ext/phar/tests/phar_metadata_write4.phpt
+ext/phar/tests/phar_oo_002.phpt
+ext/phar/tests/phar_oo_005.phpt
+ext/phar/tests/phar_oo_007.phpt
+ext/readline/tests/libedit_callback_handler_install_001.phpt
+ext/readline/tests/libedit_callback_handler_remove_001.phpt
+ext/simplexml/tests/bug51615.phpt
+ext/spl/tests/bug40091.phpt
+ext/spl/tests/bug44144.phpt
+ext/spl/tests/bug48493.phpt
+ext/spl/tests/bug62978.phpt
+ext/spl/tests/bug65006.phpt
+ext/spl/tests/bug71236.phpt
+ext/spl/tests/bug75049.phpt
+ext/spl/tests/fixedarray_003.phpt
+ext/spl/tests/spl_autoload_002.phpt
+ext/spl/tests/spl_autoload_004.phpt
+ext/spl/tests/spl_autoload_006.phpt
+ext/spl/tests/spl_autoload_008.phpt
+ext/spl/tests/spl_autoload_013.phpt
+ext/standard/tests/array/array_map_variation4.phpt
+ext/standard/tests/array/array_reverse_variation4.phpt
+ext/standard/tests/array/array_unique_variation3.phpt
+ext/standard/tests/file/disk_free_space_basic.phpt
+ext/standard/tests/file/lstat_stat_variation10.phpt
+ext/standard/tests/filters/bug54350.phpt
+ext/standard/tests/general_functions/bug73973.phpt
+ext/standard/tests/general_functions/floatval.phpt
+ext/standard/tests/general_functions/floatval_variation1.phpt
+ext/standard/tests/general_functions/get_included_files.phpt
+ext/standard/tests/general_functions/gettype_settype_basic.phpt
+ext/standard/tests/general_functions/print_r.phpt
+ext/standard/tests/general_functions/print_r_64bit.phpt
+ext/standard/tests/general_functions/proc_open_null.phpt
+ext/standard/tests/general_functions/proc_open_redirect.phpt
+ext/standard/tests/general_functions/strval.phpt
+ext/standard/tests/general_functions/type.phpt
+ext/standard/tests/general_functions/var_dump_64bit.phpt
+ext/standard/tests/math/base_convert_variation1.phpt
+ext/standard/tests/math/bindec_variation1_64bit.phpt
+ext/standard/tests/math/hexdec_variation1_64bit.phpt
+ext/standard/tests/math/octdec_variation1.phpt
+ext/standard/tests/serialize/005.phpt
+ext/standard/tests/streams/bug78902.phpt
+ext/standard/tests/streams/proc_open_bug69900.phpt
+ext/standard/tests/streams/stream_context_tcp_nodelay_fopen.phpt
+ext/standard/tests/strings/implode1.phpt
+ext/standard/tests/strings/sprintf_variation54.phpt
+sapi/cli/tests/upload_2G.phpt
+sapi/fpm/tests/log-bm-in-shutdown-fn.phpt
+sapi/fpm/tests/log-bm-limit-1024-msg-80.phpt
+sapi/fpm/tests/log-bm-limit-2048-msg-4000.phpt
+sapi/fpm/tests/log-bwd-limit-1050-msg-2048.phpt
+sapi/fpm/tests/log-bwd-limit-1050-msg-2900.phpt
+sapi/fpm/tests/log-bwd-limit-8000-msg-4096.phpt
+sapi/fpm/tests/log-bwd-msg-with-nl.phpt
+sapi/fpm/tests/log-bwp-limit-1024-msg-120.phpt
+sapi/fpm/tests/log-bwp-limit-1500-msg-3300.phpt
+sapi/fpm/tests/log-bwp-msg-flush-split-fallback.phpt
+sapi/fpm/tests/log-bwp-msg-flush-split-real.phpt
+sapi/fpm/tests/log-dwd-limit-1050-msg-2048.phpt
+sapi/fpm/tests/log-dwd-limit-1050-msg-2900.phpt
+sapi/fpm/tests/log-dwd-limit-8000-msg-4096.phpt
+sapi/fpm/tests/log-dwp-limit-1000-msg-2000.phpt
+sapi/phpdbg/tests/exceptions_003.phpt
+sapi/phpdbg/tests/stepping_001.phpt
+tests/lang/bug45392.phpt
+tests/output/ob_start_basic_002.phpt
