+++ package.xml
@@ -90,7 +90,7 @@ ${notes}
         <required>
             <php>
                 <min>7.0</min>
-                <max>8.1.99</max>
+                <max>8.2.99</max>
             </php>
             <pearinstaller>
                 <min>1.4.0</min>
