+++ tests/opcache/resolver_jit.phpt
@@ -9,13 +9,13 @@ opcache.jit_buffer_size=256M
 
 DDTrace\hook_method('BaseClass', 'speak',
     function ($This, $scope, $args) {
-        echo "${scope}::speak hooked in BaseClass.\n";
+        echo "{$scope}::speak hooked in BaseClass.\n";
     }
 );
 
 DDTrace\hook_method('ChildClass', 'speak',
     function ($This, $scope, $args) {
-        echo "${scope}::speak hooked in ChildClass.\n";
+        echo "{$scope}::speak hooked in ChildClass.\n";
     }
 );
