+++ tooling/bin/generate-final-artifact.sh
@@ -30,7 +30,7 @@ for architecture in "${architectures[@]}"; do
     tmp_folder_final_gnu_trace=$tmp_folder_final_gnu/dd-library-php/trace
     tmp_folder_final_musl_trace=$tmp_folder_final_musl/dd-library-php/trace
 
-    php_apis=(20151012 20160303 20170718 20180731 20190902 20200930 20210902)
+    php_apis=(20151012 20160303 20170718 20180731 20190902 20200930 20210902 20220829)
     if [[ $architecture == "x86_64" ]]; then
       php_apis+=(20100412 20121113 20131106)
     fi
@@ -60,7 +60,7 @@ for architecture in "${architectures[@]}"; do
     tar -xf $tmp_folder_profiling_archive -C $tmp_folder_profiling
 
     # Extension
-    php_apis=(20160303 20170718 20180731 20190902 20200930 20210902)
+    php_apis=(20160303 20170718 20180731 20190902 20200930 20210902 20220829)
     for version in "${php_apis[@]}"
     do
         mkdir -v -p \
@@ -105,7 +105,7 @@ for architecture in "${architectures[@]}"; do
         tar -xf "$tmp_folder_appsec_archive" -C $tmp_folder_appsec
 
         # Extensions
-        php_apis=(20151012 20160303 20170718 20180731 20190902 20200930 20210902);
+        php_apis=(20151012 20160303 20170718 20180731 20190902 20200930 20210902 20220829);
         for php_api in "${php_apis[@]}"; do
             mkdir -p \
                 ${tmp_folder_final_gnu_appsec}/ext/$php_api \
