+++ zend_abstract_interface/interceptor/CMakeLists.txt
@@ -1,5 +1,10 @@
-add_library(zai_interceptor "${PHP_VERSION_DIRECTORY}/interceptor.c"
-                            "${PHP_VERSION_DIRECTORY}/resolver.c")
+if(PHP_VERSION_ID LESS "80200" AND PHP_VERSION_ID GREATER "79999")
+  add_library(zai_interceptor "${PHP_VERSION_DIRECTORY}/interceptor.c"
+                              "${PHP_VERSION_DIRECTORY}/resolver_pre-8_2.c")
+else()
+  add_library(zai_interceptor "${PHP_VERSION_DIRECTORY}/interceptor.c"
+                              "${PHP_VERSION_DIRECTORY}/resolver.c")
+endif()
 
 target_include_directories(zai_interceptor PUBLIC
                                        $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}>
