+++ zend_abstract_interface/interceptor/php8/interceptor.h
@@ -1,7 +1,9 @@
 #ifndef ZAI_INTERCEPTOR_H
 #define ZAI_INTERCEPTOR_H
 
+#if PHP_VERSION_ID < 80200
 extern int zai_registered_observers;
+#endif
 
 void zai_interceptor_minit(void);
 void zai_interceptor_startup(void);
